requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
dateparser==1.1.8
et-xmlfile==1.1.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
multidict==6.0.4
numpy==1.26.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
patsy==0.5.3
psycopg2-binary==2.9.7
pycryptodome==3.19.0
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
scipy==1.11.2
six==1.16.0
SQLAlchemy==2.0.21
statsmodels==0.14.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.5
websockets==11.0.3
yarl==1.9.2
aiogram~=3.1.1